chore: cleanup all the tRPC routes
