+++ apps/blog/src/api/routes/view.ts
@@ -10,18 +10,9 @@ export const viewRouter = router({
     .input(trackViewSchemaDto)
     .output(z.void())
     .mutation(async ({ input: { postSlug }, ctx: { db, req } }) => {
-      const headersList = req.headers;
-      const ipAddress =
-        headersList.get("x-forwarded-for") ??
-        headersList.get("x-real-ip") ??
-        "127.0.0.1";
-      const userAgent = headersList.get("user-agent") ?? "unknown";
-      const viewService = new ViewService({ db });
-
-      await viewService.trackView({
-        postSlug,
-        ipAddress,
-        userAgent,
-      });
+      return await new ViewService({
+        db,
+        req,
+      }).trackView({ postSlug });
     }),
 });
