+++ apps/blog/src/api/services/view/index.ts
@@ -1,28 +1,28 @@
 import { createHash } from "crypto";
+import type { NextRequest } from "next/server";
 
 import type { DatabaseClient } from "@ashgw/db";
 import { env } from "@ashgw/env";
 import { InternalError, logger } from "@ashgw/observability";
 
-interface ViewTrackingData {
-  postSlug: string;
-  ipAddress: string;
-  userAgent: string;
-}
-
 export class ViewService {
   private readonly db: DatabaseClient;
+  private readonly req: NextRequest;
   private static readonly DEDUPLICATION_WINDOW_MS = 24 * 60 * 60 * 1000; // 24 hours
 
-  constructor({ db }: { db: DatabaseClient }) {
+  constructor({ db, req }: { db: DatabaseClient; req: NextRequest }) {
     this.db = db;
+    this.req = req;
   }
 
-  public async trackView({
-    postSlug,
-    ipAddress,
-    userAgent,
-  }: ViewTrackingData): Promise<void> {
+  public async trackView({ postSlug }: { postSlug: string }): Promise<void> {
+    const headersList = this.req.headers;
+    const ipAddress =
+      headersList.get("x-forwarded-for") ??
+      headersList.get("x-real-ip") ??
+      "127.0.0.1";
+    const userAgent = headersList.get("user-agent") ?? "unknown";
+
     try {
       // Generate fingerprint that uniquely identifies this view
       const fingerprint = this._generateFingerprint({
@@ -69,7 +69,11 @@ export class ViewService {
     postSlug,
     ipAddress,
     userAgent,
-  }: ViewTrackingData): string {
+  }: {
+    postSlug: string;
+    ipAddress: string;
+    userAgent: string;
+  }): string {
     // Hash the IP address with salt first for extra security
     const hashedIp = createHash("sha256")
       .update(ipAddress + env.IP_HASH_SALT)
